Move all bookmark doc helpers out of imports
- starts off `src/bookmarks` module entrypoint
