+++ src/background.js
@@ -11,7 +11,8 @@ import { generatePageDocId } from 'src/page-storage'
 import * as index from 'src/search/search-index'
 import db from 'src/pouchdb'
 import { generateVisitDocId } from 'src/activity-logger'
-import { generateBookmarkDocId, transformToBookmarkDoc } from 'src/imports'
+import { generateBookmarkDocId } from 'src/bookmarks'
+import { transformToBookmarkDoc } from 'src/imports'
 
 export const dataConvertTimeKey = 'data-conversion-timestamp'
 export const OVERVIEW_URL = '/overview/overview.html'
